Window ERROR: Window loading error
Last night I could not figure out why I could not get a basic window to
display with pygame. I used homebrew to install the necessary packages
but still even with the most basic window/screen display tests, I could
not get the issue to resolve. I then went on to read that there have
been many issues with pygame/python3.6/macOS to which I went ahead and
reinstalled python3 and pygame and used a different install process and
everything works as expected. I now have some of the basic features of
this galaga remake created such as a settings file to add future
application changes and also a ship class that loads in a bitmap image
and displays the content on its screen.

09/01/2020