Created bullets (Finished)
I created bullets that are fired from the ship object and are then
deleted once they leave the screen. I went ahead and built in a bullet
capacity so that a ship can only shoot a 3 bullets at a time before
reloading. I grouped the bullets into a Sprite.Group() so that I could
easily keep track of all of the bullet objects in the application. I
have also been refactoring my code constantly during this time to make
the source code more readable.

09/01/2020