New graphics (ship, aliens) added
I figured out how to branch my project and I have been solving many of
the graphical issues I had before the branch in this new testing branch.
I wanted to change the items (ship, aliens) to .png images so that they
don't have a background color outside of the image. My next goal is to
change the background from black to a space related image and see how
that changes things.

09/02/2020