Graphical changes & future improvements
I went ahead and changed the background of the game from a black screen
to an image of a galaxy. I then had to change the color of the bullet
objects from black to red. I then implemented a 'Play' button that
starts off the game and will allow a player to restart the game after
they have lost (4 lives used up). I have not added the functionality of
the 'Play' button yet but I have made it so that it renders on the
screen.
- Future idea is to add astroids that come out of the top of the screen
and move fast across the screen towards the ship.

09/03/2020