ERROR FIXED: Start screen
The issue with the start screen was that the game continued to play even
while the user had not clicked 'Play' yet. I modified the alien.py class
method update() to accept 'stats' so that I could check if the game had
started yet. This enables the alien fleet to only update when the game
has started. I also had to make this parameter change in the
update_aliens() method in the game_functions file.
- I am not sure this is the absolute best fix I have for this issue due
to the fact that I don't know if this structure would scale well.. Every
new item I add would have to have this check in the update() method.

09/03/2020