High Score added
I added a way to add a high score to the game. I still have to save the
score internally but as of right now at least I can display a high score
and have a check to test if the current score is greater than the high
score and if so, modify the high_score attribute.

09/04/2020