Project Wrapping up
I finished the vast majority of this application and I'm going to stop
here and work on a few other ideas. I want to continue improving this
game as time goes on and just have it be a constant work of mine.
POSSIBLE ISSUES:
- When a player loses a round, does the speed of the alien fleet
continue to increase?

EXTRAS:
- Add astroids
- ammo packs that give the rocketship automatic firing
instead of single round shots.
- Improve alien fleet movement
- Change fleet flying movement based on rounds
- Maybe attempt to make the game appear to be moving through space

09/04/2020